family day

+++ UI/HR/Penalty/FamilyDayPub.aspx.cs
@@ -21,7 +21,7 @@ namespace UI.HR.Penalty
         {
             try
             {
-                filePathForXML = Server.MapPath("~/HR/Penalty/Data/FD_" + Enroll + ".xml");
+                filePathForXML = Server.MapPath("~/HR/Penalty/Data/FD_" + HttpContext.Current.Session[SessionParams.USER_ID].ToString() + ".xml");
                 if (!IsPostBack)
                 {
                     pnlUpperControl.DataBind(); BindPickandDrop();
@@ -38,13 +38,13 @@ namespace UI.HR.Penalty
                     }
                 }
             }
-            catch (Exception ex){ }
+            catch { }
         }
         private void BindPickandDrop()
         {
             try
             {
-                dt = new DataTable(); try { File.Delete(filePathForXML); } catch (Exception ex){ }
+                dt = new DataTable(); try { File.Delete(filePathForXML); } catch { }
                 dt = pnlty.GetPickDropList();
                 ddlPnD.DataSource = dt; ddlPnD.DataTextField = "Names";
                 ddlPnD.DataValueField = "ID"; ddlPnD.DataBind();
@@ -54,10 +54,10 @@ namespace UI.HR.Penalty
                 ddlChild.SelectedValue = "0"; txtChild.Text = ""; ddlCGender.SelectedValue = "M"; txtCDOB.Text = "";
                 txtChild.Enabled = false; ddlCGender.Enabled = false; txtCDOB.Enabled = false; btnAdd.Enabled = false;
                 hdnconfirm.Value = "0"; dgvfml.DataSource = ""; dgvfml.DataBind(); hdnsearch.Value = ""; hdncode.Value = "";
-                txtJobstation.Text = ""; txtEmployeeSearch.Text = ""; txtDepartment.Text = "";
-                txtDesignation.Text = ""; txtJobtype.Text = ""; txtUnit.Text = "";
+                txtJobstation.Text = ""; txtEmployeeSearch.Text = ""; txtDepartment.Text = ""; hdnsdob.Value = "";
+                txtDesignation.Text = ""; txtJobtype.Text = ""; txtUnit.Text = ""; hdncdob.Value = "";
             }
-            catch (Exception ex){ }
+            catch { }
         }
         [WebMethod]
         public static List<string> GetAutoCompleteData(string strSearchKey)
@@ -87,25 +87,13 @@ namespace UI.HR.Penalty
         }
         protected void ddlPtype_SelectedIndexChanged(object sender, EventArgs e)
         {
-            if (ddlPtype.SelectedValue.ToString() == "0")
-            {
-                txtSpouse.Enabled = false; ddlSGender.Enabled = false; txtSDOB.Enabled = false;
-            }
-            else
-            {
-                txtSpouse.Enabled = true; ddlSGender.Enabled = true; txtSDOB.Enabled = true;
-            }
+            if (ddlPtype.SelectedValue.ToString() == "0") { txtSpouse.Enabled = false; ddlSGender.Enabled = false; txtSDOB.Enabled = false; }
+            else { txtSpouse.Enabled = true; ddlSGender.Enabled = true; txtSDOB.Enabled = true; }
         }
         protected void ddlChild_SelectedIndexChanged(object sender, EventArgs e)
         {
-            if (ddlChild.SelectedValue.ToString() == "0")
-            {
-                txtChild.Enabled = false; ddlCGender.Enabled = false; txtCDOB.Enabled = false; btnAdd.Enabled = false;
-            }
-            else
-            {
-                txtChild.Enabled = true; ddlCGender.Enabled = true; txtCDOB.Enabled = true; btnAdd.Enabled = true;
-            }
+            if (ddlChild.SelectedValue.ToString() == "0") { txtChild.Enabled = false; ddlCGender.Enabled = false; txtCDOB.Enabled = false; btnAdd.Enabled = false; }
+            else { txtChild.Enabled = true; ddlCGender.Enabled = true; txtCDOB.Enabled = true; btnAdd.Enabled = true; }
         }
         protected void btnAdd_Click(object sender, EventArgs e)
         {
@@ -113,21 +101,11 @@ namespace UI.HR.Penalty
             {
                 if (hdnconfirm.Value == "1")
                 {
-                    string childnm = txtChild.Text;
-                    string childgndr = ddlCGender.SelectedItem.ToString();
-                    string childdob = txtCDOB.Text;
-                    if (!DateTime.TryParse(childdob, out DateTime childDateTime))
-                    {
-                        Alert("Please insert Child Date time in correct format yyyy-MM-dd");
-                        return;
-                    }
-
-                    CreateXml(childnm, childgndr, childDateTime.ToString("yyyy-MM-dd"), "CHILD");
-                    //if (childnm.Length <= 0 || childdob.Length <= 0) { ScriptManager.RegisterStartupScript(Page, typeof(Page), "StartupScript", "alert('Please fillup child information properly.');", true); }
-                    //else { CreateXml(childnm, childgndr, childdob, "CHILD"); }
+                    string childnm = txtChild.Text; string childgndr = ddlCGender.SelectedItem.ToString(); string childdob = hdncdob.Value;
+                    CreateXml(childnm, childgndr, childdob, "CHILD");
                 }
             }
-            catch (Exception ex){ }
+            catch { }
         }
         private void CreateXml(string childnm, string childgndr, string childdob, string who)
         {
@@ -183,7 +161,7 @@ namespace UI.HR.Penalty
                 else { dgvfml.DataSource = ""; }
                 dgvfml.DataBind(); txtChild.Text = ""; txtCDOB.Text = "";
             }
-            catch (Exception ex){ }
+            catch { }
         }
         protected void dgvfml_RowDeleting(object sender, GridViewDeleteEventArgs e)
         {
@@ -193,13 +171,10 @@ namespace UI.HR.Penalty
                 dsGrid.Tables[0].Rows[dgvfml.Rows[e.RowIndex].DataItemIndex].Delete();
                 dsGrid.WriteXml(filePathForXML);
                 DataSet dsGridAfterDelete = (DataSet)dgvfml.DataSource;
-                if (dsGridAfterDelete.Tables[0].Rows.Count <= 0)
-                {
-                    File.Delete(filePathForXML); dgvfml.DataSource = ""; dgvfml.DataBind();
-                }
+                if (dsGridAfterDelete.Tables[0].Rows.Count <= 0) { File.Delete(filePathForXML); dgvfml.DataSource = ""; dgvfml.DataBind(); }
                 else { LoadGridwithXml(); }
             }
-            catch (Exception ex){ }
+            catch { }
         }
         protected void btnSubmit_Click(object sender, EventArgs e)
         {
@@ -208,28 +183,21 @@ namespace UI.HR.Penalty
                 if (hdnconfirm.Value == "1")
                 {
                     XmlDocument doc = new XmlDocument(); dt = new DataTable(); int actionby = int.Parse(Session[SessionParams.USER_ID].ToString());
-                    string pnd = ddlPnD.SelectedValue.ToString();
-                    string ptype = ddlPtype.SelectedItem.ToString();
-                    string sname = txtSpouse.Text;
-                    string sgndr = ddlSGender.SelectedItem.ToString();
-                    string sdob = txtSDOB.Text;
-                    if(!DateTime.TryParse(sdob,out DateTime spouseDateTime))
-                    {
-                        Alert("Please insert Spouse Date time in correct format yyyy-MM-dd");
-                        return;
-                    }
+                    string pnd = ddlPnD.SelectedValue.ToString(); string ptype = ddlPtype.SelectedItem.ToString();
+                    string sname = txtSpouse.Text; string sgndr = ddlSGender.SelectedItem.ToString(); string sdob = hdnsdob.Value;
                     try
                     {
                         doc.Load(filePathForXML); XmlNode dSftTm = doc.SelectSingleNode("FamilyDay"); xmlString = dSftTm.InnerXml;
                         xmlString = "<FamilyDay>" + xmlString + "</FamilyDay>";
                     }
                     catch { xmlString = ""; }
-                    dt = pnlty.Familydayinformation(0, hdncode.Value, int.Parse(pnd), ptype, sname, sgndr, spouseDateTime.ToString("yyyy-MM-dd"), actionby, xmlString);
+                    dt = pnlty.Familydayinformation(0, hdncode.Value, int.Parse(pnd), ptype, sname, sgndr, sdob, actionby, xmlString);
+                    try { File.Delete(filePathForXML); } catch { }
                     ScriptManager.RegisterStartupScript(Page, typeof(Page), "StartupScript", "alert('" + dt.Rows[0]["OutMessage"].ToString() + "');", true);
                     BindPickandDrop();
                 }
             }
-            catch (Exception ex){ }
+            catch { }
         }
     }
 }
